requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.47
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
